+++ lib/src/ui/components/ui_popupmenu.dart
@@ -26,7 +26,7 @@ enum UiPopupMenuPlacement { left, right, top, bottom }
 
 typedef HideFn = void Function(Function hideFn);
 
-class UIPopupMenu extends StatefulWidget {
+class UiPopupMenu extends StatefulWidget {
   final Widget menu;
   final ValueChanged<bool> onChange;
   final WidgetBuilder menuBuilder;
@@ -37,7 +37,7 @@ class UIPopupMenu extends StatefulWidget {
   final bool show;
   final HideFn? hideFn;
 
-  const UIPopupMenu({
+  const UiPopupMenu({
     super.key,
     required this.menu,
     required this.onChange,
@@ -53,10 +53,10 @@ class UIPopupMenu extends StatefulWidget {
 
   @override
   // ignore: library_private_types_in_public_api
-  _UIPopupMenuState createState() => _UIPopupMenuState();
+  _UiPopupMenuState createState() => _UiPopupMenuState();
 }
 
-class _UIPopupMenuState extends State<UIPopupMenu>
+class _UiPopupMenuState extends State<UiPopupMenu>
     with SingleTickerProviderStateMixin {
   late GlobalKey _key;
   bool isMenuOpen = false;
@@ -152,7 +152,7 @@ class _UIPopupMenuState extends State<UIPopupMenu>
             width: MediaQuery.of(context).size.width,
             color:
                 widget.backdrop
-                    ? context.kolor.onSurface.withAlpha(12)
+                    ? context.kolors.onSurface.withAlpha(12)
                     : Colors.transparent,
           ),
         );